improve: move generate routes
Closer to all the functions it calls.
